reorganized wrapper structure
