+++ backend/backend/Cargo.toml
@@ -59,6 +59,7 @@ dotenvy = "0.15.7"
 derive_more = { version = "1.0.0-beta.6", features = ["display", "error"] }
 regex = { version = "1.10.3", features = ["std", "perf", "unicode"] }
 itertools = "0.12.1"
+bitflags = {  version = "2.4.2", features = ["std"] }
 
 # Future Utils
 futures-util = "0.3.30"
