+++ backend/backend/src/wrapper/entity/mod.rs
@@ -1,0 +1,6 @@
+pub mod account;
+pub mod budget;
+pub mod currency;
+pub mod session;
+pub mod transaction;
+pub mod user;
