+++ backend/backend/src/wrapper/mod.rs
@@ -1,9 +1,4 @@
-pub mod account;
-pub mod budget;
-pub mod currency;
+pub mod entity;
 pub mod permission;
-pub mod session;
-pub mod transaction;
 pub mod types;
-pub mod user;
 pub mod util;
